split api functionality, add helper.py
