+++ .gitignore
@@ -1,5 +1,4 @@
-# Learn more https://docs.github.com/en/get-started/getting-started-with-git/ignoring-files
-
+# TODO: cleanup
 # dependencies
 node_modules/
 
@@ -36,3 +35,6 @@ yarn-error.*
 *.tsbuildinfo
 
 app-example
+
+__pycache__/
+*.pyc
